A bunch of rearrangement and form factor handling.
Moved a bunch of stuff out to global values (in lieu of making things
objects - this is a stepping stone, despite being a step back in the
broader sense).

Also changing the handling of form factors, to support specifying a set
of form factors and matching departure angles which are then
interpolated between to get the actual form factor.