Update match_form_factor code.
This was more involved than others, since it requires updating the form
factor tables during the calculation. It could be done more elegantly, I
think, but it works.

Also updated the form factor table output, to make it easier to cut and
paste the results into a command line that will use them. This needs a
more sophisticated appraoch (primarily a configuration file of some
form), but for now it'll do.