Lots of cleanups.
Changed format of the form factor calculation results (it's cleaner and
easier to read now, as well as easier to cut and paste for use
elsewhere); changed form for the projectile configuration information
(it now includes all the form factor data that's available, or none if
there's none (which is a step forward)).

Also fixing the range table code to match all the other changes made so
far.