Massive changes to the command processing.
This change moves all the commands and command line handling into a
collection of classes, and change all the output functions to produce
formatted text rather than print directly. A collection of CLI classes
then implement the current command line functionality, running the
analyses and printing the output.

The point of all this is that the analyses and the formatting of the
output is now separated from the command line code, and should now be
usable from a GUI program as well.

The next step will be to move all that code out to separate files and
make it a little cleaner.