Add support for building an 'empty' projectile.
This is a bit roundabout - instead of making an empty Projectile object
directly, we're providing a way to build an 'empty' set of args which
can then be used to build the Projectile.

There are better ways to do this, I'm sure, but given the complexities
of the argument/config file processing this is the least intrusive way
that I can see.