More significant changes.
I keep forgetting to commit in between implementing features . . .

The first one is proper support in the GUI for specifying the form
factor from a file. This was very fiddly, since it's using a single
control for two significantly different things (this only makes sense
because the two different things are mutually exclusive). After the
fiddling it all seems to have worked out okay.

The second feature is something approximating the foundation of an undo
stack. This was also very fiddly, since the processing model we're using
doesn't have any natural "commit these changes" point - whenever the
user runs an analysis the current state of the projectile config panel
gets slurped up, converted to a projectile object, and handed to the
analysis code. That slurping up point is the best option for the
"commit" point, but since we have no way of knowing if the contents of
the projectile config panel have changed it's a challenge to determine
if we should be adding a new undo point. To figure that out we're doing
a comparison between the top of the undo stack and the new projectile,
using a painful and fragile manual attribute by attribute comparison.

To make this work really well we need to add more information to the
undo stack - things like the time it was pushed, and possibly some way
to view the state of each undo point so that the user isn't flying
blind. Oh, and a way to actually tell the program to undo . . .