Add support for persistent configuration/status.
This makes use of a .meb directory in your home directory, with a
config.ini and status.ini file which record persistent configuration,
and persistent state. The configuration file is user managed, and the
status file is entirely programmatically managed (though user readable
and editable, as a standard ini file).

Currently the only information that's recorded is recent file history
and the last directory (both to make opening existing files easier). No
configuration is actually used yet, but the infrastructure is now in
place to support it.