Big change to form factor workflow.
After attempting to actually /use/ this thing, I've made major changes
to the form factor workflow. The big thing is keeping track of the shots
that you've specified - this is, after all, the source data you're
matching, and will be persistent across multiple other changes to the
projectile configuration. This results in something that's /much/ more
usable.

I'm considering adding this source data to the projectile file (if
available), so that you don't have to re-enter it all the time . . . But
that's a future thing.