Formatting and lint fixes.
This is making use of black for formatting. Some effort has been made to
clear all warnings, but in a few cases there wasn't any clean way to do
it, including one warning about cyclomatic complexity and a string of
warnings about 'l' being a bad variable name - this is the variable used
almost everywhere in the original code (and many other contexts) for the
departure angle, so it makes sense to stick with it here.